requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
Pillow==10.1.0
imagehash==4.3.1
python-dotenv==1.0.0
pyyaml==6.0.1
tqdm==4.66.1
fake-useragent==1.4.0
lxml==4.9.3
